fix: separate speaker modal
