+++ src/components/pages/agenda/agenda-table/speaker-modal/index.js
@@ -1,0 +1,3 @@
+import SpeakerModal from './speaker-modal';
+
+export default SpeakerModal;
